pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.1",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.10",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "uvicorn>=0.41.0",
]
